Added more classes for different components. Found error with memory limitation. This still needs to be resolved but is in a state that it will compile.

+++ StratoHawk/PayloadMain/StratoGPS.h
@@ -12,12 +12,12 @@ class StGPS{
 
     //methods
     void readStream();
-    long getLat();
-    long getLong();
+    float getLat();
+    float getLong();
 
   private:
-    long longitude;
-    long latitude;
+    float longitude;
+    float latitude;
     int RxPin, TxPin;
     uint32_t GPSBaud;
     bool isInit;
